requirements.txt
azure-identity>=1.15.0
openai>=1.12.0
python-dotenv>=1.0.0
